requirements.txt
# Django y dependencias principales
Django>=4.2,<5.0
django-extensions>=3.2.0

# Formularios y UI
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# Reportes y documentos
reportlab>=4.0.0
openpyxl>=3.1.0
Pillow>=10.0.0

# Base de datos (opcional para producción)
psycopg2-binary>=2.9.0  # Para PostgreSQL
# mysqlclient>=2.2.0     # Para MySQL

# Utilidades de desarrollo
ipython>=8.0.0
django-debug-toolbar>=4.0.0

# Servidor de producción
gunicorn>=21.0.0
whitenoise>=6.5.0

# Variables de entorno
python-decouple>=3.8

# Tareas programadas (para reportes automáticos)
celery>=5.3.0
redis>=4.6.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
coverage>=7.3.0